refactor + clean up
